+++ Sources/SuperwallKit/Config/Models/FeatureFlags.swift
@@ -12,7 +12,6 @@ struct RawFeatureFlag: Decodable {
   let enabled: Bool
 }
 
-
 struct FeatureFlags: Decodable {
   var enableSessionEvents: Bool
   var enablePostback: Bool