+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 
 ### Enhancements
 
-- Adds `useCachedTemplates` to `PaywallOptions`, which defaults to `true`. When you save a change to your paywall in the Superwall dashboard, a key is appended to the end of your paywall website URL, e.g. `sw_cache_key=<Date saved>`. This is used to cache your paywall webpage to disk after it has first loaded. Superwall will continue to load the cached version of your paywall webpage until the next time you make a change on the Superwall dashboard. With this enhancement, your paywalls will load lightning fast and will reduce network load of your app.
+- Adds paywall caching. By default, all paywalls are now cached (unless they've been updated). However, you'll be able to turn this off on the dashboard. With this enhancement, your paywalls will load lightning fast and will reduce network load of your app.
 - Exposes `Logging` `SuperwallOption` to Objective C.
 - Exposes `info` on the `PaywallViewController`.